Addition of Moodle navigation features- first pass
This adds the Navigational features to navigate between current and archive moodle in the Kent Course Overview block as well as adding the Kent Course List block, which is a block that overrides the functionality/output of the Course List block without modifying any core code.  See our Kent.Docs folder for details about adding bits to the config file to get this working.